js实现重定向拦截
